# agent/openapi.yml
openapi: 3.0.1
info:
  title: Agent API
  description: API for interacting with Agent
  version: 'v1'

paths:
  /tasks:
    post:
      summary: Creates a new task for the agent to execute.
      tags: [tasks]
      operationId: createAgentTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentTask'
      responses:
        204:
          description: The update was successful
        default:
          description: Internal Server Error
    get:
      summary: List all tasks that have been created for the agent.
      tags: [tasks]
      operationId: listAgentTasks
      responses:
        200:
          description: Successfuly listed agent tasks
          content:
            application/json:
              schema:
                type: array
                description: List of tasks IDs
                items:
                  $ref: '#/components/schemas/AgentTask'
        default:
          description: Internal Server Error

  /tasks/{task_id}/step:
    post:
      summary: Execute step in the specified task.
      tags: [tasks]
      operationId: postAgentTaskStep
      parameters:
        - name: task_id
          in: path
          description: ID of the task
          required: true
          schema:
            type: string
      responses:
        200:
          description: The step was successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStepResult'
        default:
          description: Internal Server Error

components:
  schemas:
    AgentTaskInput:
      description: Input parameters for the task. This can be any JSON serializable object.
      type: object

    AgentTask:
      type: object
      description: Wrapper containing the task.
      properties:
        task_id:
          description: The ID of the task.
          type: string
        input:
          $ref: "#/components/schemas/AgentTaskInput"

    AgentStepDeliverables:
      description: Deliverables that the task has produced.
      type: object

    AgentStepOutput:
      description: Output of the task. This can be any JSON serializable object.
      type: object

    AgentStepResult:
      allOf:
        - $ref: '#/components/schemas/AgentTask'
        - type: object
          properties:
            output:
              $ref: "#/components/schemas/AgentStepOutput"
            deliverables:
              $ref: "#/components/schemas/AgentStepDeliverables"