# agent/python/e2b_agent/models.py
# generated by fastapi-codegen:
#   filename:  ../openapi.yml
#   timestamp: 2023-07-13T15:24:06+00:00

from __future__ import annotations

from typing import Optional

from pydantic import BaseModel, Field


class AgentTaskInput(BaseModel):
    pass


class AgentTask(BaseModel):
    task_id: Optional[str] = Field(None, description='The ID of the task.')
    input: Optional[AgentTaskInput] = None


class AgentStepDeliverables(BaseModel):
    pass


class AgentStepOutput(BaseModel):
    pass


class AgentStepResult(AgentTask):
    output: Optional[AgentStepOutput] = None
    deliverables: Optional[AgentStepDeliverables] = None