# agent/python/e2b_agent/routers/tasks.py
# generated by fastapi-codegen:
#   filename:  openapi.yml
#   timestamp: 2023-07-13T15:24:06+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['tasks'])


@router.post('/tasks', response_model=None, tags=['tasks'])
def create_agent_task(body: AgentTask) -> None:
    """
    Creates a new task for the agent to execute.
    """
    pass


@router.get('/tasks', response_model=List[AgentTask], tags=['tasks'])
def list_agent_tasks() -> List[AgentTask]:
    """
    List all tasks that have been created for the agent.
    """
    pass


@router.post('/tasks/{task_id}/step', response_model=AgentStepResult, tags=['tasks'])
def post_agent_task_step(task_id: str) -> AgentStepResult:
    """
    Execute step in the specified task.
    """
    pass